Change value type to float

+++ database/greyscale.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Integer, func, ForeignKey
+from sqlalchemy import Column, DateTime, Float, Integer, func, ForeignKey
 from sqlalchemy.orm import relationship, backref
 
 from database import Base
@@ -9,7 +9,7 @@ class GreyscaleModel(Base):
     __tablename__ = 'greyscale'
     id = Column(Integer, primary_key=True, autoincrement=True)
     timestamp = Column(DateTime, default=func.now())
-    value = Column(Integer)
+    value = Column(Float)
 
     cup_id = Column(Integer, ForeignKey('cup.id'))
     cup = relationship(CupModel, uselist=False, backref=backref('greyscale_levels', uselist=True, cascade='delete,all'))
