+++ database/liquid.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, func
+from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, func
 from sqlalchemy.orm import relationship, backref
 
 from database import Base
@@ -9,7 +9,7 @@ class LiquidModel(Base):
     __tablename__ = 'liquid'
     id = Column(Integer, primary_key=True, autoincrement=True)
     timestamp = Column(DateTime, default=func.now())
-    value = Column(Integer)
+    value = Column(Float)
 
     cup_id = Column(Integer, ForeignKey('cup.id'))
     cup = relationship(CupModel, uselist=False, backref=backref('liquid_levels', uselist=True, cascade='delete,all'))
